build: add XLEN support for 32-bit userland
- XLEN={32|64} selects -march/-mabi, per-XLEN build dirs (build-32,
  build-64), link libgcc for the 64-bit divide helpers on rv32
- user.ld.S: overflow assertion is XLEN-aware
- syscall.h: register-width wrappers (long instead of long long) so the
  inline-asm register constraints hold on rv32
- wc.c: %lu -> %zu for size_t
- .gitignore: ignore the per-XLEN build directories

+++ include/sys/syscall.h
@@ -374,77 +374,74 @@
 	__asm__ __volatile__("ecall\n\t" : "=r"(a0) : __VA_ARGS__ : "memory"); \
 	return a0;
 
-static inline long long __syscall0(long long n)
+static inline long __syscall0(long n)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0");
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0");
 	__asm_syscall("r"(a7))
 }
 
-static inline long long __syscall1(long long n, long long a)
+static inline long __syscall1(long n, long a)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
 	__asm_syscall("r"(a7), "0"(a0))
 }
 
-static inline long long __syscall2(long long n, long long a, long long b)
+static inline long __syscall2(long n, long a, long b)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
-	register long long a1 __asm__("a1") = b;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
+	register long a1 __asm__("a1") = b;
 	__asm_syscall("r"(a7), "0"(a0), "r"(a1))
 }
 
-static inline long long __syscall3(long long n, long long a, long long b,
-				   long long c)
+static inline long __syscall3(long n, long a, long b, long long c)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
-	register long long a1 __asm__("a1") = b;
-	register long long a2 __asm__("a2") = c;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
+	register long a1 __asm__("a1") = b;
+	register long a2 __asm__("a2") = c;
 	__asm_syscall("r"(a7), "0"(a0), "r"(a1), "r"(a2))
 }
 
-static inline long long __syscall4(long long n, long long a, long long b,
-				   long long c, long long d)
+static inline long __syscall4(long n, long a, long b, long long c, long long d)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
-	register long long a1 __asm__("a1") = b;
-	register long long a2 __asm__("a2") = c;
-	register long long a3 __asm__("a3") = d;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
+	register long a1 __asm__("a1") = b;
+	register long a2 __asm__("a2") = c;
+	register long a3 __asm__("a3") = d;
 	__asm_syscall("r"(a7), "0"(a0), "r"(a1), "r"(a2), "r"(a3))
 }
 
-static inline long long __syscall5(long long n, long long a, long long b,
-				   long long c, long long d, long long e)
+static inline long __syscall5(long n, long a, long b, long long c, long long d,
+			      long long e)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
-	register long long a1 __asm__("a1") = b;
-	register long long a2 __asm__("a2") = c;
-	register long long a3 __asm__("a3") = d;
-	register long long a4 __asm__("a4") = e;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
+	register long a1 __asm__("a1") = b;
+	register long a2 __asm__("a2") = c;
+	register long a3 __asm__("a3") = d;
+	register long a4 __asm__("a4") = e;
 	__asm_syscall("r"(a7), "0"(a0), "r"(a1), "r"(a2), "r"(a3), "r"(a4))
 }
 
-static inline long long __syscall6(long long n, long long a, long long b,
-				   long long c, long long d, long long e,
-				   long long f)
+static inline long __syscall6(long n, long a, long b, long long c, long long d,
+			      long long e, long long f)
 {
-	register long long a7 __asm__("a7") = n;
-	register long long a0 __asm__("a0") = a;
-	register long long a1 __asm__("a1") = b;
-	register long long a2 __asm__("a2") = c;
-	register long long a3 __asm__("a3") = d;
-	register long long a4 __asm__("a4") = e;
-	register long long a5 __asm__("a5") = f;
+	register long a7 __asm__("a7") = n;
+	register long a0 __asm__("a0") = a;
+	register long a1 __asm__("a1") = b;
+	register long a2 __asm__("a2") = c;
+	register long a3 __asm__("a3") = d;
+	register long a4 __asm__("a4") = e;
+	register long a5 __asm__("a5") = f;
 	__asm_syscall("r"(a7), "0"(a0), "r"(a1), "r"(a2), "r"(a3), "r"(a4),
 		      "r"(a5))
 }
 
-#define __scc(X) ((long long)(X))
+#define __scc(X) ((long)(X))
 
 #define __syscall1(n, a) __syscall1(n, __scc(a))
 #define __syscall2(n, a, b) __syscall2(n, __scc(a), __scc(b))
